# swagger/OpenAPI/gist.rest.swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "https",
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "title": "Gist-service API.",
    "version": "0.0.1"
  },
  "host": "localhost:8082",
  "basePath": "/api/gist",
  "paths": {
...
    "/v1/{username}/forked": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "# Get Forked Gists",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Fork"
        ],
        "summary": "Get Forked Gists",
        "operationId": "get_forked_gists",
        "parameters": [
          {
            "name": "username",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "GistRequest",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/GistRequest"
              }
            }
          },
          "400": {
            "description": ""
          }
        }
      }
    },
...
    "/v1/{username}/{gist_id}/comment": {
      "get": {
        "description": "# Get Comments Of Gist",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Comment"
        ],
        "summary": "Get Comments Of Gist",
        "operationId": "get_comments",
        "parameters": [
          {
            "name": "username",
            "in": "path"
          },
          {
            "name": "gist_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Comment",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Comment"
              }
            }
          },
          "400": {
            "description": ""
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "# Create Comment",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Comment"
        ],
        "summary": "Create Comment",
        "operationId": "comment_create",
        "parameters": [
          {
            "name": "Comment",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Comment"
            }
          },
          {
            "name": "username",
            "in": "path"
          },
          {
            "name": "gist_id",
            "in": "path"
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          }
        }
      }
    },
    "/v1/{username}/{gist_id}/comment/{comment_id}": {
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "# Delete Comment By ID, if it user's comment",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Comment"
        ],
        "summary": "Delete Comment By ID",
        "operationId": "delete_comment_by_id",
        "parameters": [
          {
            "name": "username",
            "in": "path"
          },
          {
            "name": "gist_id",
            "in": "path"
          },
          {
            "name": "comment_id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          }
        }
      },
      "patch": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "# Update Comment By ID, if it is user's comment",
        "consumes": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Comment"
        ],
        "summary": "Update Comment By ID",
        "operationId": "update_comment_by_id",
        "parameters": [
          {
            "name": "Comment",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Comment"
            }
          },
          {
            "name": "username",
            "in": "path"
          },
          {
            "name": "gist_id",
            "in": "path"
          },
          {
            "name": "comment_id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          }
        }
      }
    },
    "/v1/{username}/{gist_id}/fork": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "# Fork Gist",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Fork"
        ],
        "summary": "Fork Gist",
        "operationId": "fork_gist",
        "parameters": [
          {
            "name": "username",
            "in": "path"
          },
          {
            "name": "gist_id",
            "in": "path"
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          }
        }
      }
    },
...
  },
  "definitions": {
    "Comment": {
      "type": "object",
      "properties": {
        "created_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedAt"
        },
        "gist_id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "GistID"
        },
        "id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "ID"
        },
        "text": {
          "type": "string",
          "x-go-name": "Text"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "UpdatedAt"
        },
        "user_id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "UserID"
        },
        "username": {
          "type": "string",
          "x-go-name": "Username"
        }
      },
      "x-go-package": "github.com/nanmenkaimak/github-gist/internal/gist/entity"
    },
    "Commit": {
      "type": "object",
      "properties": {
        "comment": {
          "type": "string",
          "x-go-name": "Comment"
        },
        "created_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedAt"
        },
        "gist_id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "GistID"
        },
        "id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "ID"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "UpdatedAt"
        }
      },
      "x-go-package": "github.com/nanmenkaimak/github-gist/internal/gist/entity"
    },
...
    "File": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "x-go-name": "Code"
        },
        "commit_id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "CommitID"
        },
        "created_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedAt"
        },
        "id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "UpdatedAt"
        }
      },
      "x-go-package": "github.com/nanmenkaimak/github-gist/internal/gist/entity"
    },
...
    "Gist": {
      "type": "object",
      "properties": {
        "created_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedAt"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "ID"
        },
        "is_forked": {
          "type": "boolean",
          "x-go-name": "IsForked"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "UpdatedAt"
        },
        "user_id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "UserID"
        },
        "visible": {
          "type": "boolean",
          "x-go-name": "Visible"
        }
      },
      "x-go-package": "github.com/nanmenkaimak/github-gist/internal/gist/entity"
    },
    "GistRequest": {
      "type": "object",
      "properties": {
        "commit": {
          "$ref": "#/definitions/Commit"
        },
        "files": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/File"
          },
          "x-go-name": "Files"
        },
        "gist": {
          "$ref": "#/definitions/Gist"
        }
      },
      "x-go-package": "github.com/nanmenkaimak/github-gist/internal/gist/entity"
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
...
}